Fix POS tagging pipeline
